# Improve corners matching in ChessBoardDetector::NeighborsFinder::findCornerNeighbor